% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{pumble_add_users}
\alias{pumble_add_users}
\title{Add users to a channel}
\usage{
pumble_add_users(channel_id, user_ids)
}
\arguments{
\item{channel_id}{The channel ID}

\item{user_ids}{Character vector of user IDs}
}
\description{
Add users to a channel
}
