% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{pumble_create_channel}
\alias{pumble_create_channel}
\title{Create a new channel}
\usage{
pumble_create_channel(name, type = "public", description = NULL)
}
\arguments{
\item{name}{Channel name}

\item{type}{"public" or "private"}

\item{description}{Optional description}
}
\description{
Create a new channel
}
