% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{pumble_dm_user}
\alias{pumble_dm_user}
\title{Send a direct message to a user}
\usage{
pumble_dm_user(text, user_id = NULL, email = NULL)
}
\arguments{
\item{text}{Message content}

\item{user_id}{User ID}

\item{email}{User email (if ID not provided)}
}
\description{
Send a direct message to a user
}
