% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{pumble_list_messages}
\alias{pumble_list_messages}
\title{Fetch message history in a channel}
\usage{
pumble_list_messages(channel_id = NULL, channel_name = NULL, limit = 20)
}
\arguments{
\item{channel_id}{Optional ID}

\item{channel_name}{Optional Name (if ID not provided)}

\item{limit}{Default is 20}
}
\description{
Fetch message history in a channel
}
