% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled.R
\name{pumble_schedule_message}
\alias{pumble_schedule_message}
\title{Schedule a message}
\usage{
pumble_schedule_message(channel_id, text, send_at)
}
\arguments{
\item{channel_id}{Optional ID}

\item{text}{Message content}

\item{send_at}{ISO8601 timestamp or POSIXct object}
}
\description{
Schedule a message
}
