% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{pumble_send_message}
\alias{pumble_send_message}
\title{Send a message to a channel}
\usage{
pumble_send_message(
  text,
  channel_id = NULL,
  channel_name = NULL,
  as_bot = FALSE
)
}
\arguments{
\item{text}{Message content}

\item{channel_id}{Optional ID}

\item{channel_name}{Optional Name (if ID not provided)}

\item{as_bot}{Logical, send as bot?}
}
\description{
Send a message to a channel
}
