% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{pumble_set_status}
\alias{pumble_set_status}
\title{Update custom status}
\usage{
pumble_set_status(status, emoji = NULL)
}
\arguments{
\item{status}{Text status}

\item{emoji}{Emoji code (e.g., ":joy:")}
}
\description{
Update custom status
}
